Safari-609.1.11.1/LayoutTests/inspector/animation/tracking-expected.txt

Tests that Animation.startTracking and Animation.stopTracking trigger Animation.trackingStart, Animation.trackingUpdate, and Animation.trackingComplete events with expected data.


== Running test suite: Animation.Tracking
-- Running test case: Animation.Tracking.StartAndStopTrackingWithEvent
Animation.trackingStart
PASS: Should have a timestamp when starting.
Animation.trackingUpdate
PASS: Should should have a timestamp number.
PASS: Should have an event object.
PASS: Event should have an animation tracking ID.
PASS: Event should have an animation state.
PASS: Event should have a node ID.
PASS: Event should have the related CSS animation's name.
Animation.trackingComplete
